requirements.txt
streamlit
pandas
altair
st-click-detector
openpyxl
google-generativeai
pillow
st-gsheets-connection